Download tasks now initiated if auto submit has been specified by the user. Previously completed results were uploaded but task information and forms were not synchronised.

+++ fieldTask/src/main/java/org/smap/smapTask/android/tasks/DownloadTasksTask.java
@@ -281,7 +281,8 @@ public class DownloadTasksTask extends AsyncTask<Void, String, HashMap<String, S
                 if(count > 0) {
                     results.put("Rejected Tasks", count + " deleted");
                 }
-                if(tr.settings.ft_delete_submitted) {
+
+                if(tr.settings !=null && tr.settings.ft_delete_submitted) {
                     count = Utilities.deleteTasksWithStatus(Utilities.STATUS_T_SUBMITTED);
                     if(count > 0) {
                         results.put("Submitted Tasks", count + " deleted");
